Switch to list/hash sync method
Previously, the sync button would hash all roms for a given core at
once. This has proven unworkably slow and error-prone for users with a
large number of games. As of this commit, we instead list all the games
first, without hashing anything. Subsequently, when a game is navigated
to, we show a loading spinner and get the hash for that specific game on
demand.