Remove reboot
For some unknown reason, the linux command "reboot now" doesn't seem to work in the latest version of MiSTer. So it seems better to remove it for now than to leave it in broken, and I haven't found another way that works. Also, judging by soe of Sorg's comments here and there, this might not be necessary anyway.