Remove support for custom paths
The custom paths support was creating a lot of complexity (and requiring
the user to look at a lot of popups). Some of our users would probably
prefer to retain this feature, but our feeling is that the trade off
isn't great. Our approach going forward will be to require a file
hierarchy that matches the defaults coming from a clean MiSTer fusion
setup.